stat: give st_mode, st_nlink, st_uid and st_gid their POSIX types
struct stat stored st_mode, st_nlink, st_uid and st_gid as uint16_t, while
mode_t, nlink_t, uid_t and gid_t are all 32-bit.  Code that takes the address
of a field and passes it on as the POSIX type writes four bytes into a
two-byte slot.  mc 4.8.33 does exactly that in lib/vfs/parse_ls_vga.c:

    if (!vfs_parse_filetype (p, &skipped, &s->st_mode))

with vfs_parse_filetype (..., mode_t *ret_type), which would overwrite
st_nlink on every directory listing it parses.  The 16-bit fields also cap
uids, gids and link counts at 65535 although the kernel carries 32 bits.

Declare the four fields with their POSIX types in both the userland and the
kernel <sys/stat.h>.  This is an ABI change: the struct grows from 96 to 104
bytes and every field after st_mode moves; offsets measured with the substrate
cross gcc are identical from both headers (mode 12, nlink 16, uid 20, gid 24,
rdev 28, size 32, blocks 48, atim 56, mtim 72, ctim 88, size 104).  The
kernel fills and copies out the native struct by name and by sizeof, so it
needs no code change; the personality translators assign by name into their
own foreign layouts and are unaffected.  Every substrate binary that uses
stat must be rebuilt together with this kernel.

Also updated in step:
- the _Static_assert pinning the userland size (96 -> 104);
- the private copies of the layout: bin/native_test's standalone struct,
  linux/runner's packed substrate_stat, and the native mocks in the
  tests/sys compat host tests;
- docs/specs/abi-i386.md and abi-x86_64.md tables, and stat(2);
- bootstrap.yml's toolchain cache key, v2 -> v3.  The key hashes only the
  toolchain's own inputs, but the cached libstdc++.so.6 imports stat, fstat
  and lstat compiled against the old header and would read every field past
  st_mode at the wrong offset.

Verified: make -C sys clean && make -C sys (-Werror), make -C lib and
make -C bin, make -C tests (the ci.yml suite) all pass; linux/runner builds.
The three tests/sys compat host tests do not build, but they fail identically
at HEAD (RTSIG_QUEUE_MAX, sys_munmap, USER_STACK_MIN, ...), which is
pre-existing drift unrelated to this layout, and CI does not run them.

+++ linux/runner/substrate-run.c
@@ -271,10 +271,10 @@ struct tls_image {
 struct substrate_stat {
     uint32_t st_dev;
     uint64_t st_ino;
-    uint16_t st_mode;
-    uint16_t st_nlink;
-    uint16_t st_uid;
-    uint16_t st_gid;
+    uint32_t st_mode;
+    uint32_t st_nlink;
+    uint32_t st_uid;
+    uint32_t st_gid;
     uint32_t st_rdev;
     int64_t st_size;
     uint32_t st_blksize;
