+++ tests/sys/host_test_compat_lseek32.c
@@ -42,8 +42,8 @@ typedef int32_t tid_t;
 struct stat {
     uint32_t st_dev;
     uint32_t st_ino;
-    uint16_t st_mode;
-    uint16_t st_nlink;
+    uint32_t st_mode;
+    uint32_t st_nlink;
     uint32_t st_uid;
     uint32_t st_gid;
     uint32_t st_rdev;
